refactor config logic+buttons into its own class

+++ cute_sway_recorder/main.py
@@ -1,23 +1,15 @@
 #!/usr/bin/env python3
-#
-import json
-import random
-import re
+
+from pathlib import Path
 import signal
-import string
 import subprocess
 import sys
-from datetime import datetime
-from pathlib import Path
 from subprocess import DEVNULL
-from typing import Optional
-from PySide6.QtCore import Qt
+from typing import Union
 
+from PySide6.QtCore import Qt
 from PySide6.QtWidgets import (
     QApplication,
-    QCheckBox,
-    QFileDialog,
-    QGridLayout,
     QGroupBox,
     QHBoxLayout,
     QLabel,
@@ -29,61 +21,14 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-from cute_sway_recorder.screen_selection import ScreenSelectionDialog
-
-PATTERN_FILE_WITH_SUFFIX = re.compile(r".*\..*")
-
-
-def available_screens():
-    """
-    Returns a list of all available outputs via `swaymsg -t get_outputs`, e.g.:
-    ['eDP-1', 'HDMI-A-1']
-    """
-    out = subprocess.check_output(["swaymsg", "-t", "get_outputs"], text=True)
-    return [o["name"] for o in json.loads(out)]
-
-
-def set_buttons_state(*btns, enabled: bool):
-    for b in btns:
-        b.setEnabled(enabled)
-
-
-def shrink_home(path: str) -> str:
-    """
-    Removes the tilde from path
-    /home/user/x -> ~/x
-    """
-    return path.replace(str(Path.home()), "~")
-
-
-def select_area() -> Optional[str]:
-    """
-    Launch slurp to capture a region of the screen, returns its output in the following format:
-    <x>,<y> <width>x<height>
-
-    If slurp is cancelled (by hitting escape), returns None
-    """
-    cmd = ["slurp", "-f", "%x,%y %wx%h"]
-    proc = subprocess.run(cmd, capture_output=True)
-    if proc.returncode == 0:
-        return proc.stdout.decode().strip()
-    return None
-
-
-def make_default_file_dest() -> str:
-    """
-    Create a sensible (as much as possible, you know) default name for videos that weren't given a
-    specific destination path.
-
-    Example result: /home/user/Videos/cute-NOwPn.mp4
-    """
-    identifier = "".join(random.choices(string.ascii_letters, k=5))
-    pathstr = f"~/Videos/cute-{identifier}.mp4"
-    return str(Path(pathstr).expanduser().absolute())
+from config_area import ConfigArea
+from common import SelectedArea, SelectedScreen
 
 
 def start_recording(
-    file_dst, include_audio: bool = False, area: str = None, screen: str = None
+    selection: Union[SelectedArea, SelectedScreen],
+    file_dst,
+    include_audio: bool = False,
 ) -> subprocess.Popen:
     """
     Launches a `wf-recorder` process and returns a Popen object representing
@@ -102,20 +47,17 @@ def start_recording(
 
     Saves the recording to file_dst, which is a path-like object.
     """
-    if area and screen:
-        raise ValueError("Only one of area and screen should be passed")
-
     Path(file_dst).parent.mkdir(parents=True, exist_ok=True)
 
     params = ["wf-recorder", "-f", file_dst]
     if include_audio:
         params.append("--audio")
-    if area:
+    if isinstance(selection, SelectedArea):
         params.append("--geometry")
-        params.append(area)
-    if screen:
+        params.append(selection)
+    if isinstance(selection, SelectedScreen):
         params.append("--output")
-        params.append(screen)
+        params.append(selection)
     return subprocess.Popen(params)
 
 
@@ -125,46 +67,27 @@ class CuteRecorderQtApplication:
     """
 
     def __init__(self):
-        self.selected_area = None
         self.recorder_proc = None
-        self.selected_screen = None
-        self.file_dest = make_default_file_dest()
-        self.is_whole_screen_selected = False
 
         self.app = QApplication(sys.argv)
         self.app.setApplicationDisplayName("Cute Sway Recorder")
         self.app.setDesktopFileName("cute-sway-recorder")
 
         ## Create labels
-        self.lbl_selected_area = QLabel("Selected area: None")
-        self.lbl_file_dst = QLabel(f"Saving as: {shrink_home(self.file_dest)}")
         self.lbl_is_recording = QLabel("Not recording")
-        self.lbl_whole_screen_notice = QLabel(
-            '<font color="salmon">This window will be minimized. <br/>'
-            "Click the tray icon to stop recording</font>"
-        )
-        self.lbl_whole_screen_notice.hide()
 
         ## Create buttons
-        self.btn_select_area = QPushButton("Select an area")
-        self.btn_select_whole_screen = QPushButton("Select whole screen")
         self.btn_start_recording = QPushButton("Start recording")
         self.btn_stop_recording = QPushButton("Stop recording")
-        self.btn_pick_dest = QPushButton("Pick file destination")
-        self.btn_generate_random_dest = QPushButton("Random name")
-        self.checkbox_use_audio = QCheckBox("Record audio")
 
         ## Connect buttons on-click actions
-        self.btn_select_area.clicked.connect(self.btn_onclick_select_area)
-        self.btn_select_whole_screen.clicked.connect(self.btn_onclick_select_whole_screen)
         self.btn_start_recording.clicked.connect(self.btn_onclick_start_recording)
         self.btn_stop_recording.clicked.connect(self.btn_onclick_stop_recording)
         self.btn_stop_recording.setEnabled(False)
-        self.btn_pick_dest.clicked.connect(self.btn_onclick_pick_dst)
-        self.btn_generate_random_dest.clicked.connect(self.btn_onclick_generate_random_dest)
 
         ## Show window
         self.window = QWidget()
+        self.config_area = ConfigArea(self.window)
         self.window.setWindowTitle("Cute Sway Recorder")
         self.window.setLayout(self.layout())
         self.window.show()
@@ -179,16 +102,6 @@ class CuteRecorderQtApplication:
         self.icon.activated.connect(self.tray_icon_activated_handler)
 
     def layout(self):
-        grid = QGridLayout()
-
-        grid.addWidget(self.lbl_selected_area, 0, 0)
-        grid.addWidget(self.btn_select_area, 0, 1)
-        grid.addWidget(self.btn_select_whole_screen, 0, 2)
-
-        grid.addWidget(self.lbl_file_dst, 1, 0)
-        grid.addWidget(self.btn_pick_dest, 1, 1)
-        grid.addWidget(self.btn_generate_random_dest, 1, 2)
-
         recording_btns = QHBoxLayout()
         recording_btns.addWidget(self.btn_start_recording)
         recording_btns.addWidget(self.btn_stop_recording)
@@ -200,9 +113,7 @@ class CuteRecorderQtApplication:
 
         layout = QVBoxLayout()
         layout.addWidget(lbl_recording_box)
-        layout.addWidget(self.lbl_whole_screen_notice)
-        layout.addLayout(grid)
-        layout.addWidget(self.checkbox_use_audio)
+        layout.addLayout(self.config_area)
         layout.addLayout(recording_btns)
         return layout
 
@@ -229,32 +140,17 @@ class CuteRecorderQtApplication:
             self.icon.hide()
             self.btn_onclick_stop_recording()
 
-    def btn_onclick_select_area(self):
-        self.selected_area = select_area() or self.selected_area
-        self.selected_screen = None
-        self.lbl_selected_area.setText(f"Selected area: {self.selected_area}")
-        self.is_whole_screen_selected = False
-        self.lbl_whole_screen_notice.hide()
-
-    def btn_onclick_select_whole_screen(self):
-        screens = available_screens()
-        if len(screens) > 1:
-            selected_screen_idx = ScreenSelectionDialog(screens, parent=self.window).exec()
-            if selected_screen_idx == -1:
-                return
-            self.selected_screen = screens[selected_screen_idx]
-        self.selected_area = None
-        self.lbl_selected_area.setText(f"Selected screen: {self.selected_screen}")
-        self.lbl_whole_screen_notice.show()
-        self.is_whole_screen_selected = True
-
     def btn_onclick_start_recording(self):
+        config = self.config_area.create_config()
+        if not config:
+            return
+
         # confirm dest file override
-        if Path(self.file_dest).exists():
+        if config.file_dest.exists():
             resp = QMessageBox.question(
                 self.window,
                 "File exists",
-                f"Override {self.file_dest}?",
+                f"Override {config.file_dest}?",
                 QMessageBox.Yes,
                 QMessageBox.No,
             )
@@ -262,76 +158,31 @@ class CuteRecorderQtApplication:
                 self.lbl_is_recording.setText("Not recording")
                 return
 
-        # show error message and return if no area was selected
-        if self.selected_area is None and self.selected_screen is None:
-            warning = QMessageBox(
-                QMessageBox.Critical,
-                "No area selected",
-                "Kindly select an area or pick a screen :)",
-                parent=self.window,
-            )
-            warning.exec()
-            return
-
-        # show tray icon if recording fullscreen
-        if self.is_whole_screen_selected:
+        # show tray icon if recording whole screen
+        if isinstance(config.selection, SelectedScreen):
             self.window.hide()
             self.icon.show()
 
         # disable all buttons other than stop record
         self.btn_stop_recording.setEnabled(True)
-        set_buttons_state(
-            self.btn_pick_dest,
-            self.btn_generate_random_dest,
-            self.btn_select_area,
-            self.btn_select_whole_screen,
-            self.btn_start_recording,
-            enabled=False,
-        )
-
-        # reset the "saved as" label, if it's not the first recording session
-        self.lbl_file_dst.setText(f"Saving as: {shrink_home(self.file_dest)}")
+        self.config_area.disable_all_buttons()
 
         # launch wf-recorder
         self.recorder_proc = start_recording(
-            self.file_dest,
-            include_audio=self.checkbox_use_audio.isChecked(),
-            area=self.selected_area,
-            screen=self.selected_screen,
+            config.selection,
+            config.file_dest,
+            include_audio=config.include_audio,
         )
-
-        # set respective labels
-        self.lbl_is_recording.setText('<font color="Red">RECORDING</font>')
-        self.lbl_file_dst.setText(f"Saving as: {shrink_home(self.file_dest)}")
 
     def btn_onclick_stop_recording(self):
         self.btn_stop_recording.setEnabled(False)
-        set_buttons_state(
-            self.btn_pick_dest,
-            self.btn_generate_random_dest,
-            self.btn_select_area,
-            self.btn_select_whole_screen,
-            self.btn_start_recording,
-            enabled=True,
-        )
-        self.recorder_proc.send_signal(signal.SIGINT)
-        self.lbl_is_recording.setText('<font color="Green">Done recording</font>')
-        self.lbl_file_dst.setText(
-            f'<font color="Green">Saved to: {shrink_home(self.file_dest)}</font>'
-        )
+        self.config_area.enable_all_buttons()
+        self.btn_start_recording.setEnabled(True)
 
-    def btn_onclick_pick_dst(self):
-        dest: str = QFileDialog.getSaveFileName(parent=self.window)[0]
-        if dest == "":
-            return
-        if not PATTERN_FILE_WITH_SUFFIX.match(dest):
-            dest = f"{dest}.mp4"
-        self.file_dest = dest
-        self.lbl_file_dst.setText(f"Saving as: {shrink_home(dest)}")
+        if self.recorder_proc:
+            self.recorder_proc.send_signal(signal.SIGINT)
 
-    def btn_onclick_generate_random_dest(self):
-        self.file_dest = make_default_file_dest()
-        self.lbl_file_dst.setText(f"Saving as: {shrink_home(self.file_dest)}")
+        self.lbl_is_recording.setText('<font color="Green">Saved!</font>')
 
     def exec(self):
         return self.app.exec()
